# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Appointment Scheduling API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  professional: "professional",
  client: "client",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  phone?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  admin: "admin",
  professional: "professional",
  client: "client",
} as const;

export interface CreateUserRequest {
  name: string;
  email: string;
  password: string;
  role: CreateUserRequestRole;
  phone?: string;
}

export type UpdateUserRequestRole =
  (typeof UpdateUserRequestRole)[keyof typeof UpdateUserRequestRole];

export const UpdateUserRequestRole = {
  admin: "admin",
  professional: "professional",
  client: "client",
} as const;

export interface UpdateUserRequest {
  name?: string;
  email?: string;
  phone?: string;
  role?: UpdateUserRequestRole;
}

export interface Service {
  id: number;
  name: string;
  description?: string | null;
  durationMinutes: number;
  price?: number | null;
  createdAt: string;
}

export interface CreateServiceRequest {
  name: string;
  description?: string;
  durationMinutes: number;
  price?: number;
}

export type TimeslotStatus =
  (typeof TimeslotStatus)[keyof typeof TimeslotStatus];

export const TimeslotStatus = {
  scheduled: "scheduled",
  done: "done",
  cancelled: "cancelled",
} as const;

export interface Timeslot {
  id: number;
  startTime: string;
  endTime: string;
  status: TimeslotStatus;
  notes?: string | null;
  serviceId: number;
  serviceName: string;
  professionalId: number;
  professionalName: string;
  clientId: number;
  clientName: string;
  createdAt: string;
}

export interface CreateTimeslotRequest {
  startTime: string;
  endTime: string;
  serviceId: number;
  professionalId: number;
  clientId: number;
  notes?: string;
}

export type UpdateTimeslotRequestStatus =
  (typeof UpdateTimeslotRequestStatus)[keyof typeof UpdateTimeslotRequestStatus];

export const UpdateTimeslotRequestStatus = {
  scheduled: "scheduled",
  done: "done",
  cancelled: "cancelled",
} as const;

export interface UpdateTimeslotRequest {
  startTime?: string;
  endTime?: string;
  serviceId?: number;
  professionalId?: number;
  clientId?: number;
  notes?: string;
  status?: UpdateTimeslotRequestStatus;
}

export type UpdateStatusRequestStatus =
  (typeof UpdateStatusRequestStatus)[keyof typeof UpdateStatusRequestStatus];

export const UpdateStatusRequestStatus = {
  scheduled: "scheduled",
  done: "done",
  cancelled: "cancelled",
} as const;

export interface UpdateStatusRequest {
  status: UpdateStatusRequestStatus;
}

export interface StatsOverview {
  totalScheduled: number;
  totalDone: number;
  totalCancelled: number;
  totalClients: number;
  totalProfessionals: number;
  totalServices: number;
  upcomingToday: number;
}

export type ListUsersParams = {
  role?: ListUsersRole;
};

export type ListUsersRole = (typeof ListUsersRole)[keyof typeof ListUsersRole];

export const ListUsersRole = {
  admin: "admin",
  professional: "professional",
  client: "client",
} as const;

export type ListTimeslotsParams = {
  status?: ListTimeslotsStatus;
  professionalId?: number;
  clientId?: number;
};

export type ListTimeslotsStatus =
  (typeof ListTimeslotsStatus)[keyof typeof ListTimeslotsStatus];

export const ListTimeslotsStatus = {
  scheduled: "scheduled",
  done: "done",
  cancelled: "cancelled",
} as const;