# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Appointment Scheduling API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: users
    description: User management (clients and professionals)
  - name: services
    description: Service management
  - name: timeslots
    description: Time slot scheduling
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, professional, client]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"

  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a service by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    put:
      operationId: updateService
      tags: [services]
      summary: Update a service (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "200":
          description: Service updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /timeslots:
    get:
      operationId: listTimeslots
      tags: [timeslots]
      summary: List time slots (filtered by role)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, done, cancelled]
        - name: professionalId
          in: query
          schema:
            type: integer
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Timeslot"
    post:
      operationId: createTimeslot
      tags: [timeslots]
      summary: Create a time slot (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimeslotRequest"
      responses:
        "201":
          description: Time slot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timeslot"

  /timeslots/{id}:
    get:
      operationId: getTimeslot
      tags: [timeslots]
      summary: Get a time slot by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Time slot details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timeslot"
    put:
      operationId: updateTimeslot
      tags: [timeslots]
      summary: Update a time slot (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTimeslotRequest"
      responses:
        "200":
          description: Time slot updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timeslot"
    delete:
      operationId: deleteTimeslot
      tags: [timeslots]
      summary: Delete a time slot (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Time slot deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /timeslots/{id}/status:
    patch:
      operationId: updateTimeslotStatus
      tags: [timeslots]
      summary: Update time slot status (professional only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timeslot"

  /stats/overview:
    get:
      operationId: getStatsOverview
      tags: [timeslots]
      summary: Get scheduling stats overview (admin/professional)
      responses:
        "200":
          description: Stats overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsOverview"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, professional, client]
        phone:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt

    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, professional, client]
        phone:
          type: string
      required:
        - name
        - email
        - password
        - role

    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, professional, client]

    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        durationMinutes:
          type: integer
        price:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - durationMinutes
        - createdAt

    CreateServiceRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        price:
          type: number
      required:
        - name
        - durationMinutes

    Timeslot:
      type: object
      properties:
        id:
          type: integer
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [scheduled, done, cancelled]
        notes:
          type: string
          nullable: true
        serviceId:
          type: integer
        serviceName:
          type: string
        professionalId:
          type: integer
        professionalName:
          type: string
        clientId:
          type: integer
        clientName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - startTime
        - endTime
        - status
        - serviceId
        - serviceName
        - professionalId
        - professionalName
        - clientId
        - clientName
        - createdAt

    CreateTimeslotRequest:
      type: object
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        serviceId:
          type: integer
        professionalId:
          type: integer
        clientId:
          type: integer
        notes:
          type: string
      required:
        - startTime
        - endTime
        - serviceId
        - professionalId
        - clientId

    UpdateTimeslotRequest:
      type: object
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        serviceId:
          type: integer
        professionalId:
          type: integer
        clientId:
          type: integer
        notes:
          type: string
        status:
          type: string
          enum: [scheduled, done, cancelled]

    UpdateStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [scheduled, done, cancelled]
      required:
        - status

    StatsOverview:
      type: object
      properties:
        totalScheduled:
          type: integer
        totalDone:
          type: integer
        totalCancelled:
          type: integer
        totalClients:
          type: integer
        totalProfessionals:
          type: integer
        totalServices:
          type: integer
        upcomingToday:
          type: integer
      required:
        - totalScheduled
        - totalDone
        - totalCancelled
        - totalClients
        - totalProfessionals
        - totalServices
        - upcomingToday