# lib/api-zod/src/generated/types/createServiceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Appointment Scheduling API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateServiceRequest {
  name: string;
  description?: string;
  durationMinutes: number;
  price?: number;
}