# lib/api-zod/src/generated/types/createTimeslotRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Appointment Scheduling API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTimeslotRequest {
  startTime: Date;
  endTime: Date;
  serviceId: number;
  professionalId: number;
  clientId: number;
  notes?: string;
}