# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Appointment Scheduling API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  description?: string | null;
  durationMinutes: number;
  price?: number | null;
  createdAt: Date;
}